create xa stress test class and moved related tests to there, renamed THREAD_FLAG to TRANSACTION_EXISTS
