+++ hazelcast/src/test/java/com/hazelcast/xa/HazelcastXACompatibilityTest.java
@@ -149,61 +149,4 @@ public class HazelcastXACompatibilityTest extends HazelcastTestSupport {
         xaResource.recover(XAResource.TMSTARTRSCAN | XAResource.TMENDRSCAN);
     }
 
-    @Test
-    public void testCommitConcurrently() throws InterruptedException, XAException {
-        int count = 10000;
-        String name = randomString();
-        ExecutorService executorService = Executors.newFixedThreadPool(5);
-        ExecutorService executorServiceForCommit = Executors.newFixedThreadPool(5);
-        for (int i = 0; i < count; i++) {
-            XATransactionRunnable runnable = new XATransactionRunnable(xaResource, name, executorServiceForCommit, i);
-            executorService.execute(runnable);
-        }
-        IMap<Object, Object> map = instance.getMap(name);
-        assertSizeEventually(count, map);
-    }
-
-    static class XATransactionRunnable implements Runnable {
-
-        HazelcastXAResource xaResource;
-
-        String name;
-
-        ExecutorService executorServiceForCommit;
-
-        int i;
-
-        public XATransactionRunnable(HazelcastXAResource xaResource, String name,
-                                     ExecutorService executorServiceForCommit, int i) {
-            this.xaResource = xaResource;
-            this.name = name;
-            this.executorServiceForCommit = executorServiceForCommit;
-            this.i = i;
-        }
-
-        @Override
-        public void run() {
-            try {
-                final Xid xid = createXid();
-                xaResource.start(xid, XAResource.TMNOFLAGS);
-                TransactionContext context = xaResource.getTransactionContext();
-                TransactionalMap<Object, Object> map = context.getMap(name);
-                map.put(i, i);
-                xaResource.end(xid, XAResource.TMSUCCESS);
-                executorServiceForCommit.execute(new Runnable() {
-                    @Override
-                    public void run() {
-                        try {
-                            xaResource.commit(xid, true);
-                        } catch (XAException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                });
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
 }
